start to adapt pipeline.cpp
