+++ mad_icp_ros/config/mad_icp/jackal.yaml
@@ -1,0 +1,26 @@
+mad_icp_odometry:
+  ros__parameters:
+    # mad_icp
+    b_max : 0.2 # [m] max size of kd leaves
+    b_min : 0.1 # [m] when a node is flatten than this param, propagate normal
+    b_ratio : 0.02 # the increase factor of search radius needed in data association 
+    p_th : 0.8 # [%] ensuring an update when the curr point cloud is registered less than this param
+    rho_ker : 0.1 # huber threshold in mad-icp
+    n : 10 # the number of last poses to smooth velocity
+    use_wheels : true # use wheel odometry
+    num_threads : 4 # number of threads
+
+    # sensor
+    min_range : 0.1
+    max_range : 50.0
+    sensor_hz : 10.0
+    deskew : False
+    lidar_in_base: 
+      [1.000,  0.000,  0.000, -0.120,
+      0.000,  1.000,  0.000,  0.000,
+      0.000,  0.000,  1.000,  0.584,
+      0.000,  0.000,  0.000, 1.000]
+    intensity_thr : 0.0 # filter low intensity points
+
+    # ROS2
+    base_frame : "base_link"
